requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.3
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
